adding a cursor prop for the icon so we can specify a cursor for icon elements
